refactor: split up cli into separate modules
