+++ app/src/cli/mod.rs
@@ -4,6 +4,17 @@ use clap::{Args, Parser, Subcommand};
 
 pub mod new_project;
 
+pub mod assets;
+pub mod build;
+pub mod client;
+pub mod deploy;
+pub mod server;
+
+mod project_path;
+pub use project_path::*;
+
+use self::assets::AssetCommand;
+
 #[derive(Parser, Clone)]
 #[command(author, version, about, long_about = None)]
 #[command(propagate_version = true)]
@@ -83,36 +94,6 @@ pub enum Commands {
         #[command(subcommand)]
         command: AssetCommand,
     },
-}
-
-#[derive(Args, Clone, Debug)]
-pub struct MigrateOptions {
-    #[arg(index = 1, default_value = "./assets")]
-    /// The path to the assets folder
-    pub path: PathBuf,
-}
-
-#[derive(Args, Clone, Debug)]
-pub struct ImportOptions {
-    #[arg(index = 1)]
-    /// The path to the assets you want to import
-    pub path: PathBuf,
-    #[arg(long)]
-    /// Whether to convert audio files to OGG
-    pub convert_audio: bool,
-    /// Whether to generate a collider from the model
-    #[arg(long)]
-    pub collider_from_model: bool,
-}
-
-#[derive(Subcommand, Clone, Debug)]
-pub enum AssetCommand {
-    /// Migrate json pipelines to toml
-    #[command(name = "migrate-pipelines-toml")]
-    MigratePipelinesToml(MigrateOptions),
-    /// Import new assets with interactive prompts
-    #[command(name = "import")]
-    Import(ImportOptions),
 }
 
 #[derive(Subcommand, Clone, Copy, Debug)]
